+++ tower-http/src/compression/future.rs
@@ -42,6 +42,8 @@ where
 
         // never recompress responses that are already compressed
         let should_compress = !res.headers().contains_key(header::CONTENT_ENCODING)
+            // never compress responses that are ranges
+            && !res.headers().contains_key(header::CONTENT_RANGE)
             && self.predicate.should_compress(&res);
 
         let (mut parts, body) = res.into_parts();
@@ -100,6 +102,7 @@ where
             }
         };
 
+        parts.headers.remove(header::ACCEPT_RANGES);
         parts.headers.remove(header::CONTENT_LENGTH);
 
         parts